metasploitable3-attack-scripts-master/requirements.txt
pwntools
requests
pyyaml
beautifulsoup4
fabric
sliver-py
prettytable
inquirer